requirements.txt
datasets==3.6.0
numpy==1.26.4
omegaconf==2.3.0
perch_hoplite==0.1.0
# perch-hoplite @ git+https://github.com/google-research/perch-hoplite.git@7e5ddf30b720555dc9ef0d03182ac924c94ce873
pyrootutils==1.0.4
Requests==2.32.3
scikit_activeml==0.6.2
scikit_learn==1.7.0
scipy==1.15.3
sentence_transformers==4.1.0
tensorflow==2.19.0
torch==2.7.0
tqdm==4.67.1
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
tf-keras==2.19.0
tensorflow-hub==0.16.1
